Move Cursor to its own component
