+++ src/chart/generateCategoricalChart.tsx
@@ -1,4 +1,4 @@
-import React, { Component, cloneElement, isValidElement, createElement, ReactElement } from 'react';
+import React, { Component, cloneElement, isValidElement, ReactElement } from 'react';
 import isNil from 'lodash/isNil';
 import isFunction from 'lodash/isFunction';
 import range from 'lodash/range';
@@ -12,17 +12,13 @@ import clsx from 'clsx';
 // eslint-disable-next-line no-restricted-imports
 import type { DebouncedFunc } from 'lodash';
 import invariant from 'tiny-invariant';
-import { getRadialCursorPoints } from '../util/cursor/getRadialCursorPoints';
 import { getTicks } from '../cartesian/getTicks';
 import { Surface } from '../container/Surface';
 import { Layer } from '../container/Layer';
 import { Tooltip } from '../component/Tooltip';
 import { Legend } from '../component/Legend';
-import { Curve } from '../shape/Curve';
-import { Cross } from '../shape/Cross';
-import { Sector } from '../shape/Sector';
 import { Dot } from '../shape/Dot';
-import { isInRectangle, Rectangle } from '../shape/Rectangle';
+import { isInRectangle } from '../shape/Rectangle';
 
 import {
   filterProps,
@@ -83,14 +79,14 @@ import {
   Margin,
   StackOffsetType,
   TickItem,
+  TooltipEventType,
 } from '../util/types';
 import { AccessibilityManager } from './AccessibilityManager';
 import { isDomainSpecifiedByUser } from '../util/isDomainSpecifiedByUser';
 import { getActiveShapeIndexForTooltip, isFunnel, isPie, isScatter } from '../util/ActiveShapeUtils';
 import { Props as YAxisProps } from '../cartesian/YAxis';
 import { Props as XAxisProps } from '../cartesian/XAxis';
-import { getCursorPoints } from '../util/cursor/getCursorPoints';
-import { getCursorRectangle } from '../util/cursor/getCursorRectangle';
+import { Cursor } from '../component/Cursor';
 
 export interface MousePointer {
   pageX: number;
@@ -1313,7 +1309,7 @@ export const generateCategoricalChart = ({
       this.throttleTriggeredAfterMouseMove.cancel();
     }
 
-    getTooltipEventType() {
+    getTooltipEventType(): TooltipEventType {
       const tooltipItem = findChildByType(this.props.children, Tooltip);
 
       if (tooltipItem && typeof tooltipItem.props.shared === 'boolean') {
@@ -1771,58 +1767,23 @@ export const generateCategoricalChart = ({
         this.state;
       const tooltipEventType = this.getTooltipEventType();
       // The cursor is a part of the Tooltip, and it should be shown (by default) when the Tooltip is active.
-      const isActive = element.props.active ?? isTooltipActive;
-
-      if (
-        !element ||
-        !element.props.cursor ||
-        !isActive ||
-        !activeCoordinate ||
-        (chartName !== 'ScatterChart' && tooltipEventType !== 'axis')
-      ) {
-        return null;
-      }
+      const isActive: boolean = element.props.active ?? isTooltipActive;
       const { layout } = this.props;
-      let restProps;
-      let cursorComp: React.ComponentType<any> = Curve;
 
-      if (chartName === 'ScatterChart') {
-        restProps = activeCoordinate;
-        cursorComp = Cross;
-      } else if (chartName === 'BarChart') {
-        restProps = getCursorRectangle(layout, activeCoordinate, offset, tooltipAxisBandSize);
-        cursorComp = Rectangle;
-      } else if (layout === 'radial') {
-        const { cx, cy, radius, startAngle, endAngle } = getRadialCursorPoints(activeCoordinate);
-        restProps = {
-          cx,
-          cy,
-          startAngle,
-          endAngle,
-          innerRadius: radius,
-          outerRadius: radius,
-        };
-        cursorComp = Sector;
-      } else {
-        restProps = { points: getCursorPoints(layout, activeCoordinate, offset) };
-        cursorComp = Curve;
-      }
-      const key = element.key || '_recharts-cursor';
-      const cursorProps = {
-        stroke: '#ccc',
-        pointerEvents: 'none',
-        ...offset,
-        ...restProps,
-        ...filterProps(element.props.cursor),
-        payload: activePayload,
-        payloadIndex: activeTooltipIndex,
-        key,
-        className: 'recharts-tooltip-cursor',
-      };
-
-      return isValidElement(element.props.cursor)
-        ? cloneElement(element.props.cursor, cursorProps)
-        : createElement(cursorComp, cursorProps);
+      return (
+        <Cursor
+          activeCoordinate={activeCoordinate}
+          activePayload={activePayload}
+          activeTooltipIndex={activeTooltipIndex}
+          chartName={chartName}
+          element={element}
+          isActive={isActive}
+          layout={layout}
+          offset={offset}
+          tooltipAxisBandSize={tooltipAxisBandSize}
+          tooltipEventType={tooltipEventType}
+        />
+      );
     };
 
     renderPolarAxis = (element: any, displayName: string, index: number) => {
